Insert Test Particle into a user defined slab
The test particle can be inserted into a slab along the z-axis
between tpizmin and tpizmax (values specified in the mdp-file).
If no value is given the particle will randomly placed all over the box.
If the values are equal the particle is placed at a certain
z-coordinate.

Thanks to Robert Vacha for posting some code snipplets on gmx-devel.