Make g_density average the slice-width (important for NPT)
g_density bins the density profiles into a fixed number of bins.
In the original version, the final slice-width is used for creating the
output file axis. Now the slice-width is averaged over the trajectory
such that positions of the density profile can be compared. This is
especially useful for comparing positions of cosolutes in membrane
simulations!